Enhances JSON model attribute handling
Improves the handling of JSON attributes within the CrelishJsonModel.
It ensures that updates to JSON fields are properly reflected and prevents
the creation of dynamic properties when setting JSON attributes.
Additionally, it ensures that only defined properties or attributes are populated
during model initialization from JSON data.

Updates version to 0.14.19.